Add more Add/Sub(Assign) methods to geometry
The primary motivation behind this patch was the implementation of the
`Add` and `Sub` traits for `Size` and `Point`. This is very useful when
having to reduce a size based on a location (to limit boundaries for
example).

The `AddAssign` and `SubAssign` implementations for `Size` seem to go
hand-in-hand with the normal `Add` and `Sub` implementations and since
they were missing for `Point` too I decided to just implement all
permutations.