Update render matrix for rotated outputs
When the output is rotated by 90°/270° using a transform in the
`Gles2Renderer::render` function, the resulting matrix would cause the
buffer to be stretched along one and compressed along the other axis.

Since width and height are swapped when rotating by 90 degrees
(counter-)clockwise, the resulting output matrix needs to adjust for
this difference.