Move shared input event functions to common traits
The current state of input events is somewhat of a tragedy for
consumers. You write your implementation in the match branch for the
individual event, then notice there are a lot of similarities, but after
you try to move it into a function you realize that all the input events
are completely separate events that just have methods named the same
way.

To allow downstream to actually make use of these traits as proper
shared functionality and write their own helper abstractions, new traits
have been introduced that are implemented for all events.

These two new traits, `PositionEvent` and `TouchEvent`, allow accessing
the slot and location of any generic input event which provides this
functionality.