Fix error due to unknown DRM property
This resolves a regression introduced in 6260e46, which could lead to
compositors not finding any viable GPU due to their lack of the
`rotation` property.

Since this property is only set to ensure it has a sane default value,
it should not be necessary on devices where it is not present at all.

Fixes #536.